requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt
drf-yasg==1.20.0
pyjwt==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
factory_boy